Add AuraHUD, Overlay Widget and Widget Controllers
Added AuraHUD class and OverlayWidget for UI elements. Implemented Widget Controller for controlling UI elements, such as the newly added OverlayWidget. In addition, applied necessary changes to the MainCharacter class for initializing AuraHUD and integrated this with Player Controller. The changes also introduced a callback binding in the Widget Controllers to allow changes in gameplay attributes.