Add AuraGameplayEffect and update EffectActor with gameplay effects
The commit creates a new AuraGameplayEffect class and modifies the EffectActor class to utilize gameplay effects. It also includes changing character base classes to implement the AbilitySystemInterface. This update allows for more complex gameplay interactions with the newly implemented effects on the target actors. The AuraGameplayEffect is created to be used with effects on the gameplay target.